server-python/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-dotenv>=1.0.0
agora-agent-server-sdk==1.3.0
